requirements.txt
asgiref==3.8.1
Django==5.1.5
django-stubs==5.1.2
django-stubs-ext==5.1.2
mypy==1.14.1
mypy-extensions==1.0.0
pillow==11.1.0
psycopg2-binary==2.9.10
python-dotenv==1.0.1
sqlparse==0.5.3
types-PyYAML==6.0.12.20241230
typing_extensions==4.12.2
